+++ browser extension/plugin/src/ui-components/atoms/i18n.jsx
