+++ browser extension/plugin/src/ui-components/atoms/language.js
@@ -1,3 +1,5 @@
+// todo : move to config.js
+
 export const langNameMap = {
   English: "en",
   Hindi: "hi",
