+++ browser extension/plugin/src/ui-components/pages/Api.js
@@ -1,5 +1,5 @@
 import axios from "axios";
-import config from "./config";
+import config from "../../config";
 
 const { API_URL } = config;
