+++ browser extension/plugin/src/ui-components/pages/App.jsx
