+++ browser extension/plugin/src/ui-components/pages/Preferences.jsx
