+++ browser extension/plugin/src/ui-components/pages/Resources.jsx
