+++ browser extension/plugin/src/ui-components/pages/options.jsx
